File read IO work in progress: Changed value_representation::get functions to be non-const since those functions have side effects Updated everything to be const-correct for the non-const get functions Added a vr_value_list template class to encapsulate the get/set implementation behavior Updated vr classes to use vr_value_list Preliminary implementation of seekable_stream interface to be used by "other" vrs. Will allow for data to be saved to temporary files Added a memory_stream implementation of the seekable_stream interface Updated other_vr to use a seekable_stream

+++ src/fume/file_tx_stream.cpp
@@ -33,7 +33,7 @@ file_tx_stream::file_tx_stream( const string&     filename,
     assert( m_callback != nullptr );
 }
 
-MC_STATUS file_tx_stream::write( const void* buffer, size_t buffer_bytes )
+MC_STATUS file_tx_stream::write( const void* buffer, uint32_t buffer_bytes )
 {
     MC_STATUS ret = MC_CANNOT_COMPLY;
     if( buffer != nullptr && m_callback != nullptr )
